correct code to clear rating as we need to force 0 to short so npgsql finds the right proc

+++ src/Maw.Data/PhotoRepository.cs
@@ -229,7 +229,7 @@ public class PhotoRepository
                 {
                     photoId,
                     username = username.ToLowerInvariant(),
-                    score = 0,
+                    score = (short)0,
                     roles
                 }
             );
